server_producer/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.5.14
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.4
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.1
google-api-core==2.11.1
google-auth==2.23.0
google-cloud==0.34.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.4
google-cloud-pubsublite
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
interchange==2021.0.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
langchain==0.0.294
langsmith==0.0.38
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.26.0
onnxruntime==1.15.1
openai==0.28.0
outcome==1.2.0
overrides
packaging==23.1
pansi==2020.7.3
posthog==3.0.2
proto-plus==1.22.3
protobuf==4.24.3
pulsar-client==3.3.0
py2neo==2021.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
pymongo==4.5.0
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
selenium==4.12.0
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
